Add Delete and Edit function

+++ frontend/src/Admin/Pages/Detailing/EditRealizationD/EditRealizationD.js
@@ -10,15 +10,17 @@ import { Link } from "react-router-dom";
 import { Button } from "react-bootstrap";
 import { REACT_APP_START_URL } from "../../../../environmentVariables";
 import { getRole } from "../../../../redux/commonRedux";
-const EditRealizationD = (id) => {
+
+const EditRealizationD = () => {
   const realization = useSelector(getRealization);
   const dispatch = useDispatch();
   const user = useSelector(getRole);
-  console.log(user);
 
-  const handleDelete = () => {
+  const handleDelete = (id) => {
+    if (!window.confirm("Czy na pewno chcesz usunąć tę realizację?")) return;
     dispatch(deleteRealizationD(id));
   };
+
   return (
     <div className={style.container}>
       {user === "ADMIN" ? (
@@ -50,7 +52,7 @@ const EditRealizationD = (id) => {
           ))}
         </div>
       ) : (
-        <h1>Not founfd</h1>
+        <h1>Not found</h1>
       )}
     </div>
   );
